moved classes out of base64_app to their own files
